Corrections-6
web3.js updated from 0.20.x to 1.2.x version.
Added callers into the functions with .call so it runs locally as a simulation so if it fails; no Ether is spent.
Notifications to the user that their transaction has been submitted. (on transactionHash, on receipt, on error).
Check that receipt status = true.
Get gasPrice from tx.